Psx Data Reader microservice/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
orjson>=3.10
beautifulsoup4>=4.12
lxml>=5.3
pandas>=2.2
openpyxl>=3.1
requests>=2.32
urllib3>=2.2
python-dotenv>=1.1
